Moved all the code into one file. Vex didnt like it when i split it up.
