+++ src/main.py
@@ -1,7 +1,6 @@
-
 from vex import *
-# from drive import rc_auto_loop_function_controller_1
 
+#CONFIG
 
 brain=Brain()
 
@@ -9,9 +8,140 @@ controller=Controller(PRIMARY)
 
 
 
+
+
+
+####### LIFT STUFF ↓ #######    
+
+lift = Motor(Ports.PORT5)
+
+lift.set_stopping(BrakeType.HOLD)
+
+
+def on_up_button_pressed():
+    lift.spin(direction=REVERSE)
+    print("lift up")
+
+def on_down_button_pressed():
+    lift.spin(direction=FORWARD)
+    print("lift down")
+    
+
+
+def on_lift_button_released():
+    lift.stop()
+
+
+
+# controller event handlers
+controller.buttonR1.pressed(on_up_button_pressed)
+controller.buttonR1.released(on_lift_button_released)
+controller.buttonR2.pressed(on_down_button_pressed)
+controller.buttonR2.released(on_lift_button_released)
+
+
+####### CLAW CODE ↓ ########
+
+claw = Motor(Ports.PORT12)
+
+claw.set_stopping(BrakeType.HOLD)
+
+
+def on_close_button_pressed():
+    claw.spin(direction=REVERSE)
+
+def on_open_button_pressed():
+    claw.spin(direction=FORWARD)
+    
+
+
+def on_claw_button_released():
+    claw.stop()
+
+
+
+# controller event handlers
+controller.buttonL1.pressed(on_open_button_pressed)
+controller.buttonL1.released(on_claw_button_released)
+controller.buttonL2.pressed(on_close_button_pressed)
+controller.buttonL2.released(on_claw_button_released)
+
+
+
+####### DRIVE STUFF ↓ ########
+
+
+
+# motors and drive train setup
+left_motor_a = Motor(Ports.PORT1, GearSetting.RATIO_18_1, False)
+left_motor_b = Motor(Ports.PORT2, GearSetting.RATIO_18_1, False)
+left_drive_smart = MotorGroup(left_motor_a, left_motor_b)
+right_motor_a = Motor(Ports.PORT3, GearSetting.RATIO_18_1, True)
+right_motor_b = Motor(Ports.PORT4, GearSetting.RATIO_18_1, True)
+right_drive_smart = MotorGroup(right_motor_a, right_motor_b)
+drivetrain = DriveTrain(left_drive_smart, right_drive_smart, 319.19, 355.59999999999997, 266.7, MM, 1)
+
+#varibles needed for the controoler loop
+drivetrain_l_needs_to_be_stopped_controller_1 = False
+drivetrain_r_needs_to_be_stopped_controller_1 = False
+
+
+
+# define a task that will handle monitoring inputs from controller_1
+def rc_auto_loop_function_controller_1():
+    global drivetrain_l_needs_to_be_stopped_controller_1, drivetrain_r_needs_to_be_stopped_controller_1, controller_1_left_shoulder_control_motors_stopped, controller_1_x_b_buttons_control_motors_stopped, remote_control_code_enabled
+    # process the controller input every 20 milliseconds
+    # update the motors based on the input values
+    while True:
+        if remote_control_code_enabled:
+            
+            # calculate the drivetrain motor velocities from the controller joystick axies
+            # left = axis3 + axis1
+            # right = axis3 - axis1
+            drivetrain_left_side_speed = controller.axis3.position()
+            drivetrain_right_side_speed = controller.axis2.position()
+            
+            # check if the value is inside of the deadband range
+            if drivetrain_left_side_speed < 5 and drivetrain_left_side_speed > -5:
+                # check if the left motor has already been stopped
+                if drivetrain_l_needs_to_be_stopped_controller_1:
+                    # stop the left drive motor
+                    left_drive_smart.stop()
+                    # tell the code that the left motor has been stopped
+                    drivetrain_l_needs_to_be_stopped_controller_1 = False
+            else:
+                # reset the toggle so that the deadband code knows to stop the left motor next
+                # time the input is in the deadband range
+                drivetrain_l_needs_to_be_stopped_controller_1 = True
+            # check if the value is inside of the deadband range
+            if drivetrain_right_side_speed < 5 and drivetrain_right_side_speed > -5:
+                # check if the right motor has already been stopped
+                if drivetrain_r_needs_to_be_stopped_controller_1:
+                    # stop the right drive motor
+                    right_drive_smart.stop()
+                    # tell the code that the right motor has been stopped
+                    drivetrain_r_needs_to_be_stopped_controller_1 = False
+            else:
+                # reset the toggle so that the deadband code knows to stop the right motor next
+                # time the input is in the deadband range
+                drivetrain_r_needs_to_be_stopped_controller_1 = True
+            
+            # only tell the left drive motor to spin if the values are not in the deadband range
+            if drivetrain_l_needs_to_be_stopped_controller_1:
+                left_drive_smart.set_velocity(drivetrain_left_side_speed, PERCENT)
+                left_drive_smart.spin(FORWARD)
+            # only tell the right drive motor to spin if the values are not in the deadband range
+            if drivetrain_r_needs_to_be_stopped_controller_1:
+                right_drive_smart.set_velocity(drivetrain_right_side_speed, PERCENT)
+                right_drive_smart.spin(FORWARD)
+            # check the buttonL1/buttonL2 status
+            # to control scooper_motor
+        # wait before repeating the process
+        wait(20, MSEC)
+
+
+
 # define variable for remote controller enable/disable
 remote_control_code_enabled = True
 
-# rc_auto_loop_thread_controller_1 = Thread(rc_auto_loop_function_controller_1)
-
-        
+rc_auto_loop_thread_controller_1 = Thread(rc_auto_loop_function_controller_1)
